+++ python/pyautd3/__init__.py
@@ -4,7 +4,7 @@ Project: pyautd3
 Created Date: 24/05/2021
 Author: Shun Suzuki
 -----
-Last Modified: 28/09/2023
+Last Modified: 29/09/2023
 Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 -----
 Copyright (c) 2022-2023 Shun Suzuki. All rights reserved.
@@ -47,4 +47,4 @@ __all__ = [
     "TimerStrategy",
 ]
 
-__version__ = "15.2.0"
+__version__ = "15.2.1"