+++ python/pyautd3/gain/focus.py
@@ -4,7 +4,7 @@ Project: gain
 Created Date: 14/09/2023
 Author: Shun Suzuki
 -----
-Last Modified: 14/09/2023
+Last Modified: 29/09/2023
 Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 -----
 Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -22,15 +22,31 @@ from ..internal.gain import IGain
 
 
 class Focus(IGain):
+    """Gain to produce a focal point
+
+    """
+
     _p: np.ndarray
     _amp: Optional[float]
 
     def __init__(self, pos: np.ndarray):
+        """Constructor
+
+        Arguments:
+        - `pos` - Position of the focal point
+        """
+
         super().__init__()
         self._p = pos
         self._amp = None
 
     def with_amp(self, amp: float) -> "Focus":
+        """Set amplitude
+
+        Arguments:
+        - `amp` - Normalized amplitude (from 0 to 1)
+        """
+
         self._amp = amp
         return self