+++ src/examples/src/remote_twincat.rs
@@ -4,15 +4,13 @@
  * Created Date: 22/05/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 05/09/2023
+ * Last Modified: 29/09/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
  *
  */
 
-#[macro_use]
-mod test_runner;
 mod tests;
 
 use anyhow::Result;
@@ -21,9 +19,9 @@ use autd3::prelude::*;
 use autd3_link_twincat::RemoteTwinCAT;
 
 fn main() -> Result<()> {
-    let mut autd = Controller::builder()
+    let autd = Controller::builder()
         .add_device(AUTD3::new(Vector3::zeros(), Vector3::zeros()))
         .open_with(RemoteTwinCAT::new("0.0.0.0.0.0")?)?;
 
-    run!(autd)
+    tests::run(autd)
 }