+++ src/examples/src/tests/mod.rs
@@ -4,7 +4,7 @@
  * Created Date: 28/05/2021
  * Author: Shun Suzuki
  * -----
- * Last Modified: 15/09/2023
+ * Last Modified: 29/09/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2021 Shun Suzuki. All rights reserved.
@@ -20,15 +20,7 @@ mod group;
 mod holo;
 mod plane;
 mod stm;
+mod test_runner;
 mod transtest;
 
-pub use audio_file::*;
-pub use bessel::*;
-pub use custom::*;
-pub use flag::*;
-pub use focus::*;
-pub use group::group;
-pub use holo::*;
-pub use plane::*;
-pub use stm::*;
-pub use transtest::*;
+pub use test_runner::run;